fix(s3): allow import S3 bucket with a legacy name (#30679)
### Issue # (if applicable)

Closes #22640.

### Reason for this change

Customers could not imported S3 bucket that has a legacy name (contains underscore in the name) and use it in CDK Apps.

### Description of changes

This change allowed customer to use S3 buckets legacy names for only the imported S3 buckets, but not for the new ones.

### Description of how you validated changes

Added unit test cases, and integration test cases.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*