chore: add new partitions and regions (#30737)
Note that the docstring in `aws-entities` states that the regions are
added in the order they went live. For several, that was not accurate so
I reordered them. This is unlikely to matter but I made the change so
that the documentation there would be factually correct.

Not all of these regions/partitions are live yet, but they have been
announced [here](https://aws.amazon.com/blogs/aws/category/regions/)

### Checklist
- [ ] My code adheres to the [CONTRIBUTING
GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and
[DESIGN
GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*