docs(events): improved maintenance of information on supported targets (#30785)
### Reason for this change
Information about targets supported by `aws-events-targets` is also listed in `aws-events`, but this information is outdated and can be confusing to users.
Aggregating the list of supported targets in `aws-events-targets` will prevent contributors from forgetting to update it.

### Description of changes
Add a message to the README of `aws-events` encouraging people to look at `aws-events-targets` for information on targets.

### Description of how you validated changes



### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*