feat(logs): enabling IA log group creation in CN and GovCloud regions (#30904)
### Reason for this change
CloudwatchLogs recently launched support for Infrequent Access Log Group Class in china and gov cloud regions. We are adding support in CDK to use this feature.

### Description of changes
Added LogGroup Class attribute support in CDK for China and GovCloud regions. This feature is already launched by CWL in china and gov-cloud regions. This PR would add CDK support for the same.

### Description of how you validated changes
This feature is already has integ test coverage in commercial region. We have validated the feature on service side already.

### Checklist
- [-] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*