feat(ecs): add validation checks to memory cpu combinations of FARGATE compatible task definitions (#30166)
### Issue # (if applicable)

Closes #22216

### Reason for this change

Adds a validation that would point out invalid CPU - Memory limits for a FARGATE compatible task definition. This is based on the document https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-cpu-memory-error.html

### Description of changes

Added a `validateFargateTaskDefinitionMemoryCpu` method in the base task definition and run this validation when a task definition has FARGATE in requiresCompatibilities. 

This is in line with the ECS requirement that even if a task definition has EC2 compatibility (along with FARGATE), FARGATE validations apply

### Description of how you validated changes 

Validated the changes by providing incorrect as well as correct values, fixed all the invalid combinations that are present in the unit tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*