docs(opensearchservice): fix broken Markdown due to an invisible char (#31007)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

For some invisible char in the Markdown, it wasn't rendered correctly.

### Description of changes

I am unsure what was there, but replacing the "space" between `##` and `Enable` with a regular space (i.e., ' ') fixed it.

### Description of how you validated changes

After this change, using the Markdown preview feature in my editor, as well as regular Markdown syntax highlighting, worked as expected.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*