chore: add readme for cloud-assembly-schema (#31151)
This points users to the correct repo for the source code. No other changes have yet been made to this README. The contents of this README will be automatically updated to the contents of the README in the source code on every version update.

This change also includes removal of a file that shouldn't have been committed (.jsii.tabl.json) and the revert of unintended changes to the jest config.

### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change



### Description of changes



### Description of how you validated changes



### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*