chore: save the logs
The tests for init are finding the cli's readme instead of the one it would be generating for a new app/library.
When run in the cli, it's quite nice to have the README output, when we run the tests and get the cli's entire README
it makes our build/test output unreadably long.