chore: save the logs (#31274)
The tests for init are finding the cli's readme instead of the one it would be generating for a new app/library. When run in the cli, it's quite nice to have the README output, when we run the tests and get the cli's entire README it makes our build/test output unreadably long.

This change doesn't get rid of quite all the unnecessary logs but it does decrease the log output by about 30,000 lines.


### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*