docs(elasticloadbalancingv2): update confusing AddNetworkTargets port description (#30208)
### Reason for this change

The existing `port` description can be confusing as `listener` could easily be mistaken as a reference to a Load Balancers `listener` and imply that the value should match.

Additionally, "traffic" seems more semtantically correct for an NLB than "requests".

### Description of changes

Updated the `port` description on `AddNetworkTargetsProps`  to match the one used in the `ApplicationTargetGroup` as that uses less ambigous language.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*