chore: add option unacknowledged to cdk notices (#31250)
### Issue # (if applicable)

Closes #23078

### Reason for this change

`cdk notices` will show all the notices the cdk version is affected by.
The additional command `cdk notices --unacknowledged` will show the notices the customer hasn't acknowledge with `cdk acknowledge <NUMBER>`.

The output will display the number of unacknowledged notices, with the following line as the last statement:
`There are <NUMBER> unacknowledged notice(s).`
Customer or automation tools can use this command to filter for the number of unacknowledged notices.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*