fix(pipelines): "Node with duplicate id" on duplicate stack names
When having 2 stacks with the same name in the same stage (which makes
sense when deploying them to different environments), the CodePipeline
Action name is derived from the stack name, and will be duplicated.

Detect if an graph node name is already being used and if so, use
environment information to try and make the name unique.