fix(cli): deployment errors are printed 3 times
The CLI prints deployment errors 3 times. This is caused by an catching
an error, printing it, and then throwing it again; to another `catch`
statement that catches the error, prints it, and then throws it again.

In this PR, get rid of one catch and change the error that gets rethrown
in a different case.

Also in this PR: fix the inconsistency of printing the progress of asset
publishing. Compared to the progress of stack deployments, the stack
name isn't bold and there is a single space offset.