chore(dynamodb): update waitForReplicationToFinish flag summary doc to highlight that it may cause deletion for existing replicas. (#31432)
### Description of changes

Add a warning message to the summary doc of the flag `waitForReplicationToFinish` to highlight that misusing this flag may cause deletion for existing replicas. This change is required to resolve some internal feedback as some people may miss reading the detail documentation of that flag, and so they will may misuse it and then have some serious issues.

### Description of how you validated changes
see the below image for the updated doc:

<img width="728" alt="Screenshot 2024-09-12 at 3 42 22 PM" src="https://github.com/user-attachments/assets/bc264d5a-d926-47a3-a2a7-98010f7c2c8c">


### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*