fix(cli): bootstrap respects qualifier from cdk.json (#31410)
closes #28249. 

The qualifier property can be set via the context key "@aws-cdk/core:bootstrapQualifier" and if omitted, the arbitrary default is `hnb659fds`. In the case of `cdk bootstrap`, there is an additional way to set the qualifier via the `--qualifier` CLI option. Specific to the `cdk bootstrap` logic, we currently only honor the command line argument, which is an error.

Ultimately, the following `cdk bootstrap` calls should be identical:

- `cdk bootstrap` with the following `cdk.json` file:

```json
{
  "@aws-cdk/core:bootstrapQualifier": "abcde",
}
```

- `cdk bootstrap --qualifier="abcde"`

I've made the decision that the `--qualifier` parameter takes precedent over the `cdk.json` context if they are both set.