feat(ecs-patterns): pidmode option for fargate services
`pidMode` is an option of `FargateTaskDefinition`, but this was not
exposed by `FargateTaskDefinition` which allows users to provide options
for the `FargateTaskDefinition` that is created as part of a
`*FargateService` when the `taskDefinition` property is excluded. If a
user wanted to set the `pidMode` for their fargate tasks, it would
require creating a separate `FargateTaskDefinition` construct and
passing it to the `*FargateService` pattern.

This change exposes this option directly so that users can specify a
`pidMode` for their tasks without having to construct a separate
`FargateTaskDefinition`. The associated
`runtimePlatform.operatingSystemFamily` property which is required to be
set to `ecs.OperatingSystemFamily.LINUX` when using `pidMode` was already
exposed.

Closes: #31412